+++ internal/model/master_product_model.go
@@ -6,8 +6,8 @@ type MasterProduct struct {
 	ID           int64   `gorm:"primaryKey;autoIncrement" json:"id"`
 	Name         string  `gorm:"size:255;not null" json:"name"`
 	Code         string  `gorm:"size:100;unique;not null" json:"code"`
-	Quantity     int     `json:"quantity"`
-	CurrentStock int     `json:"current_stock"`
+	Quantity     int64   `json:"quantity"`
+	CurrentStock int64   `json:"current_stock"`
 	IsUnlimited  bool    `gorm:"default:false" json:"is_unlimited"`
 	Sort         int     `gorm:"default:0" json:"sort"`
 	Icon         *string `gorm:"size:255" json:"icon,omitempty"`
